refac: MapScreen 로직, UI 분리
